Modyfikacja, wyszukiwarka. Gotowe

+++ HomeLibrary/Views/MainView.cs
@@ -16,13 +16,15 @@ namespace HomeLibrary.Views
     public partial class MainView : UserControl, Interfaces.IMainView
     {
         private IViewChanger viewChanger;
-        private List<Book> books; 
+        private List<Book> books;
+        string sorted;
         public MainView(IViewChanger viewChanger)
         {
             InitializeComponent();
             books = new List<Book>();
             this.viewChanger = viewChanger;
-            
+            listViewBooks.ColumnClick += sort;
+
         }
 
         public List<Book> Books
@@ -71,6 +73,14 @@ namespace HomeLibrary.Views
 
         private void buttonDeleteBook_Click(object sender, EventArgs e)
         {
+            if (listViewBooks.SelectedItems.Count < 1)
+            {
+                MessageBox.Show("Wybierz co chcesz usunąć.");
+                return;
+            }
+            DialogResult result = MessageBox.Show(string.Format("Napewno chcesz usunąć wybrane książki?"), "Napewno?", MessageBoxButtons.YesNo);
+            if (result == DialogResult.No)
+                return;
             foreach (ListViewItem x in listViewBooks.SelectedItems)
             {
                 if (DeleteBook(uint.Parse(x.Text)))
@@ -80,7 +90,141 @@ namespace HomeLibrary.Views
 
         private void textBoxSearchBar_TextChanged(object sender, EventArgs e)
         {
-            
+            search();
+            showBooks();           
+        }
+        private void comboBoxSearch_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            search();
+            showBooks();
+        }
+        private bool validation(string s)
+        {
+            int i;
+            if(int.TryParse(s, out i))
+            {
+                return true;
+            }
+            return false;
+        }
+
+        private void search()
+        {
+            switch (comboBoxSearch.SelectedIndex)
+            {
+                case 0:
+                    Books = SearchByAuthor(textBoxSearchBar.Text);
+                    break;
+                case 1:
+                    Books = SearchByTitle(textBoxSearchBar.Text);
+                    break;
+                case 2:
+                    if (!validation(textBoxSearchBar.Text))
+                    { Books = GetBooks(); break; }
+                    Books = SearchByYear(int.Parse(textBoxSearchBar.Text));
+                    break;
+                case 3:
+                    Books = SearchByPublisher(textBoxSearchBar.Text);
+                    break;
+            }
+        }
+
+        private void showBooks()
+        {
+            listViewBooks.Items.Clear();
+            foreach (var book in books)
+            {
+                string[] bookInfo = { book.Id.ToString(), book.Title, book.Year.ToString(), book.Author.ToString(), book.Publisher };
+                ListViewItem bookItem = new ListViewItem(bookInfo);
+                listViewBooks.Items.Add(bookItem);
+            }
+        }
+
+        private void sort(object sender, ColumnClickEventArgs e)
+        {
+            int i = e.Column;
+            switch (i) {
+                case 0:
+                    if (sorted == "ascId")
+                    {
+                        Books = books.OrderByDescending(x => x.Id).ToList();
+                        sorted = "descId";
+                    }
+                    else
+                    {
+                        Books = books.OrderBy(x => x.Id).ToList();
+                        sorted = "ascId";
+                    }
+                    break;
+                case 1:
+                    if (sorted == "ascTitle")
+                    {
+                        Books = books.OrderByDescending(x => x.Title).ToList();
+                        sorted = "descTitle";
+                    }
+                    else
+                    {
+                        Books = books.OrderBy(x => x.Title).ToList();
+                        sorted = "ascTitle";
+                    }
+                    break;
+                case 2:
+                    if (sorted == "ascYear")
+                    {
+                        Books = books.OrderByDescending(x => x.Year).ToList();
+                        sorted = "descYear";
+                    }
+                    else
+                    {
+                        Books = books.OrderBy(x => x.Year).ToList();
+                        sorted = "ascYear";
+                    }
+                    break;
+                case 3:
+                    if (sorted == "ascAuthor")
+                    {
+                        Books = books.OrderByDescending(x => x.Author.LastName).ToList();
+                        sorted = "descAuthor";
+                    }
+                    else
+                    {
+                        Books = books.OrderBy(x => x.Author.LastName).ToList();
+                        sorted = "ascAuthor";
+                    }
+                    break;
+                case 4:
+                    if (sorted == "ascPublisher")
+                    {
+                        Books = books.OrderByDescending(x => x.Publisher).ToList();
+                        sorted = "descPublisher";
+                    }
+                    else
+                    {
+                        Books = books.OrderBy(x => x.Publisher).ToList();
+                        sorted = "ascPublisher";
+                    }
+                    break;
+            }
+
+            showBooks();
+        }
+
+        private void buttonModify_Click(object sender, EventArgs e)
+        {
+            if(listViewBooks.SelectedItems.Count == 1)
+            {
+                uint? id = null;
+                foreach(  ListViewItem x in listViewBooks.SelectedItems)
+                {
+                    id = uint.Parse(x.Text);
+                }
+                if (id != null)
+                    viewChanger.ShowModifyView(books.Find(x=>x.Id == id));
+                else
+                    MessageBox.Show("Wybierz jedna ksiazke do edycji.");
+            }else MessageBox.Show("Wybierz jedna ksiazke do edycji.");
         }
     }
+
+    
 }
